+++ e2e/e2etests/test_solana_deposit_and_call_revert_with_call.go
@@ -35,18 +35,13 @@ func TestSolanaDepositAndCallRevertWithCall(r *runner.E2ERunner, args []string) 
 	require.NoError(r, err)
 
 	// create encoded msg
-	abiArgs, err := solanacontracts.GetExecuteMsgAbi()
-	require.NoError(r, err)
-	msg := solanacontracts.ExecuteMsg{
-		Accounts: []solanacontracts.AccountMeta{
-			{PublicKey: [32]byte(connectedPda.Bytes()), IsWritable: true},
-			{PublicKey: [32]byte(r.ComputePdaAddress().Bytes()), IsWritable: false},
-			{PublicKey: [32]byte(solana.SystemProgramID.Bytes()), IsWritable: false},
-		},
-		Data: data,
+	accounts := []solanacontracts.AccountMeta{
+		{PublicKey: [32]byte(connectedPda.Bytes()), IsWritable: true},
+		{PublicKey: [32]byte(r.ComputePdaAddress().Bytes()), IsWritable: false},
+		{PublicKey: [32]byte(solana.SystemProgramID.Bytes()), IsWritable: false},
 	}
 
-	msgEncoded, err := abiArgs.Pack(msg)
+	msgEncoded, err := solanacontracts.EncodeExecuteMessage(accounts, data)
 	require.NoError(r, err)
 
 	sig := r.SOLDepositAndCall(nil, reverterAddr, depositAmount, data, &solanacontracts.RevertOptions{