+++ zetaclient/compliance/compliance_test.go
@@ -24,29 +24,29 @@ func TestCctxRestricted(t *testing.T) {
 	t.Run("should return true if sender is restricted", func(t *testing.T) {
 		cfg.ComplianceConfig.RestrictedAddresses = []string{cctx.InboundParams.Sender}
 		config.SetRestrictedAddressesFromConfig(cfg)
-		require.True(t, IsCctxRestricted(cctx))
+		require.True(t, IsCCTXRestricted(cctx))
 	})
 	t.Run("should return true if receiver is restricted", func(t *testing.T) {
 		cfg.ComplianceConfig.RestrictedAddresses = []string{cctx.GetCurrentOutboundParam().Receiver}
 		config.SetRestrictedAddressesFromConfig(cfg)
-		require.True(t, IsCctxRestricted(cctx))
+		require.True(t, IsCCTXRestricted(cctx))
 	})
 	t.Run("should return false if sender and receiver are not restricted", func(t *testing.T) {
 		// restrict other address
 		cfg.ComplianceConfig.RestrictedAddresses = []string{"0x27104b8dB4aEdDb054fCed87c346C0758Ff5dFB1"}
 		config.SetRestrictedAddressesFromConfig(cfg)
-		require.False(t, IsCctxRestricted(cctx))
+		require.False(t, IsCCTXRestricted(cctx))
 	})
 	t.Run("should be able to restrict coinbase address", func(t *testing.T) {
 		cfg.ComplianceConfig.RestrictedAddresses = []string{ethcommon.Address{}.String()}
 		config.SetRestrictedAddressesFromConfig(cfg)
 		cctx.InboundParams.Sender = ethcommon.Address{}.String()
-		require.True(t, IsCctxRestricted(cctx))
+		require.True(t, IsCCTXRestricted(cctx))
 	})
 	t.Run("should ignore empty address", func(t *testing.T) {
 		cfg.ComplianceConfig.RestrictedAddresses = []string{""}
 		config.SetRestrictedAddressesFromConfig(cfg)
 		cctx.InboundParams.Sender = ""
-		require.False(t, IsCctxRestricted(cctx))
+		require.False(t, IsCCTXRestricted(cctx))
 	})
 }